refactor: Common Interface Route Change
