+++ Dev/Web/src/components/Modal/ModalButton/style.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import theme from "@/shared/theme";
-import ModalButtonProps from "@/interfaces/Modal/ModalButton/ModalButtonProps";
+import ModalButtonProps from "@/interfaces/props/ModalButtonProps.ts";
 
 export const ModalButton = styled.button<ModalButtonProps>`
   background-color: ${(props) =>
@@ -10,13 +10,13 @@ export const ModalButton = styled.button<ModalButtonProps>`
   text-align: ${(props) => props.textAlign || "center"};
   color: ${(props) => {
     if (props.color === theme.colorSystem.neutral["200"]) {
-      return theme.colorSystem.neutral["500"];
+        return theme.colorSystem.neutral["500"];
     } else if (props.color === theme.colorSystem.primary["600"]) {
-      return "#ffffff";
+        return "#ffffff";
     } else {
-      return "inherit";
+        return "inherit";
     }
-  }};
+}};
   display: flex;
   width: 100%;
   padding: 15px 0px;
